quickstart/pyproject.toml
[project]
name = "quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
